pyproject.toml
[project]
name = "systematic-crypto"
version = "0.1.0"
description = "Systematic crypto trading using Carver's framework + Simons' data rigor"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas-ta>=0.4.67b0",
    "scikit-learn>=1.4",
    "pyyaml>=6.0",
    "streamlit>=1.30",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
ml = ["lightgbm>=4.0", "xgboost>=2.0", "catboost>=1.2", "optuna>=3.5"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
